requirements.txt
pynput
python-vlc
python-dotenv